implemented checkplot to gzipped pickle fn, some remaining work
we now have a checkplot.checkplot_pickle function. this can
take in lspinfo dicts from multiple period finder methods for an
object, and generate a gzipped pickle containing the object info,
finder chart, mag series plot, and for each lspinfo dict:
a periodogram, and phased mag series plots for up to however
many 'best' peaks the user wants.

this is intended for use with the more advanced checkplot viewer
app to be implemented as a Tornado webapp: checkplotserver. this
will also enable updating the varinfo and objectinfo to set the
object's variability info, best periods and epochs, and tags based
on variability type and object type.

remaining stuff to do in checkplot.py: write a checkplot pickle
file to PNG function that replicates the functionality of
twolsp_checkplot_png for arbitrary number of period-finding methods

this will look something like:

[    finder    ] [  objectinfo  ] [ variableinfo ] [ unphased LC  ]
[ periodogram1 ] [ phased LC P1 ] [ phased LC P2 ] [ phased LC P3 ]
[ periodogram2 ] [ phased LC P1 ] [ phased LC P2 ] [ phased LC P3 ]
                                 .
                                 .
[ periodogramN ] [ phased LC P1 ] [ phased LC P2 ] [ phased LC P3 ]

for N independent period-finding methods producing:

  - periodogram1,2,3...N: the periodograms from each method
  - phased LC P1,P2,P3: the phased lightcurves using the best 3 peaks
                        in each periodogram