cpserver UI: finally fixed annoying sidebar bug
The problem was in a jquery selector that was too broad: we were
looking for all instances of an anchor in the checkplot list that
had a specific attribute, then doing an unwrap to remove bold
formatting after it had been unselected. This was unknowingly catching
the <div></div> wrap around the entry for the same object in the
reviewed objects list, killing its div and thus messing up the counter
and everything in that list.

Fixed by using a selector that makes sure to only select the class
as well as the desired attribute of the anchors we want.

I still fucking hate Javascript with the heat of a million AGN.


Former-commit-id: 751ee3a729e926e194e541ade377a472b4cad84c