add centroid analysis tools in astrokep
Two new functions:

1. `detrend_centroid`, a routine that detrends the raw Kepler centroid
timeseries (see docstring for explanation of what this means).

2. `get_centroid_offsets`. Given the output of `detrend_centroid` and
ingress and egress time of transits, this returns a dictionary with info
relevant for a centroid false positive analysis.

This commit also includes a few utility functions used in the
detrending. The variable-order Legendre series detrending has a
functional form to be attached in this commit.