varbase.features: fix annoyingly subtle bug causing np.polyfit crashes
Some measurements errors returned by fiphot, et al. in the HAT pipelines are 0.0
(where they should probably be nan). This causes various division by zero
problems whenever we do a weighted calculation. In particular: np.polyfit goes
into a hard crash and just spins forever. This was the likely the root cause of
the weird FUTEX wait issue I saw with mp.Pool. This happened with
ProcessPoolExecutor and was finally tracked down running lcproc.varfeatures in
an old-fashioned for loop.