macf: fixed, tested; reproduces McQuillian+ (2014) results on KeplerLCs
more testing is probably required. doesn't seem to work so well for light curves
that have correlated noise left over like ground-based LCs. the correlogram for
these types of LCs does not drop to zero at large lags and it's unclear which
peaks to pick out for the period calculation.