checkplotserver: added a standalone mode for serving checkplot pickles
This mode is enabled by the argument --standalone=1 and requires a path
to a secret key file in --sharedsecret argument. Only one URL handler
is brought up in this mode. This serves a requested checkplot file at
path anywhere on disk as JSON over the special /standalone HTTP
endpoint.

The use-case for these is a data server loading object info, finder
charts, phased LCs and variability info that was written to checkplots
after people looked at them. The frontend data-server gets these using
async HTTP requests to a running instance of checkplotserver
(definitely on the same localhost for safety). We use this pattern in
the HAT data server to avoid heavy loads when a result table of search
query results is returned and we want nice finders, phased LCs without
doing the work all over again.

There's rudimentary authentication enabled using a pre-shared key that
must be present in the request for it to succeed. Standalone is definitely
not meant to run anywhere except for localhost.