make rms measurement point selection more generic
I initially thought it would be best to measure the RMS in-transit when
computing a transit's SNR.

However, this noise measure can be inflated if you have a bad model fit.

Sometimes, you might instead want the out of transit points.

This commit renames things in order that any passed indices can be used.