Use correct period with periodepsilon in zgls.py
Noticed this, maybe I'm wrong but it made more sense for me that the bestperiodsdiff values be compared with the current period, not the previous period.  I haven't checked whether a similar algorithm is used in the other period search algorithms but if so and this change is valid, those should be changed too.