periodbase: add transit least squares
This commit adds a wrapper to Hippke and Heller's transit least squares
package (2019; arxiv.org/abs/1901.02015).

A few comments:

* `test_tls_parallel()` passes

* the `resultdict` returned by `tls_parallel_pfind` doesn't exactly
match the astrobase standard, because of the details of the
transitleastsquares inputs/outputs.  In particular, this wrapper
relinquishes control of the frequency grid, in favor of the one Hippke &
Heller implemented.

* a review before merging would be appreciated!