minor improvements to mandel-agol model fitting
1) the exposure time is now an explicit kwarg. this is good, because
before it was not.

2) I'm now also returning the mean acceptance fraction and an estimate
of the autocorrelation time from the MCMC.

These two numbers can be used to get some indication of how good or bad
the fit went.

Generally though, it would be better to set up these fitting functions
to automatically determine the required length needed to converge,
perhaps with a "maximum length of sampling timeout" feature added on
(https://dfm.io/posts/autocorr).  The autocorrelation time is one
heuristic to make this work. I'm currently working on something like
this for mandel-agol transit fitting.