[WIP] add fivetransitparam_fit_magseries
@waqasbhatti, if you could comment before I PR, it'd be much
appreciated.

Overall explanation: I found that often I had some light curve with
multiple transits, and I wanted to fit a transit model for (t0, period,
incl, a/rstar, rp/rstar), and "standard" default choices for everything
else.

This seems generic enough to write a wrapper to the earlier
mandelagol_fit_magseries utilities.

One question @waqasbhatti: for tests, what is the preferred mechanism
for storing data? Push the file to `astrobase-notebooks` and then
download it as in the other available tests? (And why is this the
preferred mechanism? The LCs I'll be testing on are 300kb each).