add services.convert_identifiers
This was a piece of functionality that I found I often needed: you have
some identifier, and you want an identifier from another catalog.

For `simbad2gaiadrtwo`, the solution is implemented by basically just
calling the SIMBAD name resolver. If the SIMBAD name resolver returns a
Gaia identifier, this will succeed.

For `gaiadrtwo2tic`, important context is that TICv8 was rebased on Gaia
DR2. However from what I can tell, querying TICv8 using only the Gaia ID
is rather annoyingly difficult.

Instead, what this function does is use the Gaia source_id is used to
get the position, and then the position is used to query TICv8. The
result is then searched for anything matching the original source_id.

Annoyingly complicated, but it works.